default.xml: Introduce manifest to build U-Boot
The Android R&D, which we are building should use the Android U-Boot.
That one is built with the help of Bazel and repo. Repo is based on
usage of the xml manifest.

This patch is introducing the default.xml repo manifest, which
contains links to all required projects, which are needed to build
Android U-Boot, including the u-boot fork from the xen-troops.

Signed-off-by: Vladyslav Goncharuk <vladyslav_goncharuk@epam.com>